backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
anthropic[bedrock,vertex]>=0.39.0
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1
pillow==10.1.0
asyncio-mqtt==0.16.1
httpx==0.25.2
python-dotenv==1.0.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
strenum>=0.4.15